requirements.txt
kubernetes
requests
setuptools
oauth2client
httplib2
urllib3
rsa==4.0
PyYaml
requests
requests-oauthlib
requests-toolbelt
paramiko
sqlalchemy
Werkzeug==0.14.1
flask_sqlalchemy
Flask==1.0.2
flask_request_validator==2.1.0
flask-restful-swagger
Flask-API==1.1
flask_restplus==0.12.1
functools32
jsonschema
flask_monitoringdashboard==2.1.4
decorator
plotly==2.2.0
configparser
python-memcached-stats
pymemcache
itsdangerous
python-ldap
pyrsistent==0.15.2